+++ CHANGES.rst
@@ -13,6 +13,8 @@ Changelog
   Update `serialized_obj` `review_state` key if transitions were triggered in
   `FolderPost.wf_transitions`.
   [gbastien]
+- Added endpoint `@infos` to get various informations about application.
+  [gbastien]
 
 1.0a3 (2020-06-08)
 ------------------